refactor(seeder): move the seeder code to the root of the project
