+++ seeder_test.go
@@ -6,8 +6,8 @@ import (
 	"os"
 	"testing"
 
+	"github.com/danvergara/seeder"
 	"github.com/danvergara/seeder/db/seeds"
-	"github.com/danvergara/seeder/pkg/seeder"
 	"github.com/jmoiron/sqlx"
 	"github.com/stretchr/testify/assert"
